Esta promoção leva ao `main` dois trabalhos que fecharam em `dev`: o Mustard passa a custar **um gesto** onde custava três, e o `cargo clippy` sai de **347 avisos para zero** com o CI passando a reprovar qualquer aviso novo.

+++ apps/rt/src/commands/review/review_dispatch.rs
@@ -148,7 +148,7 @@ pub fn run(opts: ReviewDispatchOpts) {
     };
     let body = serde_json::to_string_pretty(&report).unwrap_or_else(|_| "{}".to_string());
     println!("{body}");
-    economy::emit_operation(&context::cwd(), ActorKind::Orchestrator, "review-dispatch", total as u64, opts.spec.as_deref(), json!({}));
+    economy::emit_operation(&context::cwd(), ActorKind::Orchestrator, "review-dispatch", total, opts.spec.as_deref(), json!({}));
 }
